+++ lib/iam/manager/Manager.rb
@@ -1,0 +1,30 @@
+require "conf/Configuration"
+require "iam/manager/IamGroups"
+require "iam/manager/IamRoles"
+require "iam/manager/IamUsers"
+
+require "aws-sdk"
+
+module Cumulus
+  module IAM
+
+    # Public: The main class for the IAM management module.
+    class Manager
+
+      attr_reader :groups
+      attr_reader :roles
+      attr_reader :users
+
+      # Public: Constructor
+      def initialize
+        iam = Aws::IAM::Client.new(
+          region: Configuration.instance.region
+        )
+        @groups = IamGroups.new(iam)
+        @roles = IamRoles.new(iam)
+        @users = IamUsers.new(iam)
+      end
+
+    end
+  end
+end
