+++ lib/route53/loader/Loader.rb
@@ -3,16 +3,20 @@ require "conf/Configuration"
 require "route53/models/ZoneConfig"
 
 # Public: Load Route53 assets
-module Loader
-  include BaseLoader
+module Cumulus
+  module Route53
+    module Loader
+      include Common::BaseLoader
 
-  @@zones_dir = Configuration.instance.route53.zones_directory
+      @@zones_dir = Configuration.instance.route53.zones_directory
 
-  # Public: Load all the zone configurations as ZoneConfig objects
-  #
-  # Returns an array of ZoneConfig
-  def self.zones
-    BaseLoader.resources(@@zones_dir, &ZoneConfig.method(:new))
+      # Public: Load all the zone configurations as ZoneConfig objects
+      #
+      # Returns an array of ZoneConfig
+      def self.zones
+        Common::BaseLoader::resources(@@zones_dir, &ZoneConfig.method(:new))
+      end
+
+    end
   end
-
 end
